Add cardinal_3 to FMapFacts
cardinal_3 asserts that updating a key already present in a map does not
change the number of keys in the map. This complements cardinal_2, which
makes the similar statement for if the key is not present.

It uses a new lemma, Add_transpose_neqkey, which states that it is
possible to transpose two map updates provided they are for different
keys. Since of course `add x e (add y f m) <> add y f (add x e m)`
necessarily, even if `x <> y`, this uses `Add` instead.